documentation: (Toy) move canonicalization patterns into dialect definition
